fixup! Add global context
