+++ src/snowflake/cli/api/commands/flags.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import tempfile
+from dataclasses import dataclass
 from enum import Enum
 from inspect import signature
 from pathlib import Path
@@ -532,3 +533,25 @@ def deprecated_flag_callback_enum(msg: str):
         return value.value
 
     return _warning_callback
+
+
+@dataclass
+class Variable:
+    key: str
+    value: str
+
+    def __init__(self, key: str, value: str):
+        self.key = key
+        self.value = value
+
+
+def parse_key_value_variables(variables: List[str]) -> List[Variable]:
+    """Util for parsing key=value input. Useful for commands accepting multiple input options."""
+    result = []
+    for p in variables:
+        if "=" not in p:
+            raise ClickException(f"Invalid variable: '{p}'")
+
+        key, value = p.split("=", 1)
+        result.append(Variable(key.strip(), value.strip()))
+    return result
